% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_map.R
\name{create_map_plot}
\alias{create_map_plot}
\title{Create Map Plot}
\description{
Creates a ggplot map from calculated indicator result.
}
\usage{
create_map_plot(indicator_map, params)
}
\arguments{
\item{indicator_map}{Result from calc_indicator_map().}
\item{params}{List of plotting parameters from create_map_plot_params().}
}
\value{
ggplot object.
}
\examples{
\dontrun{
map_result <- calc_indicator_map(...)
params <- create_map_plot_params(map_result, input, parsed)
plot <- create_map_plot(map_result, params)
}
}
