% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_map_config.R
\name{create_map_plot_params}
\alias{create_map_plot_params}
\title{Create Map Visualization Parameters}
\description{
Builds a parameter list for b3gbi map plotting functions from UI inputs.
}
\usage{
create_map_plot_params(indicator_result, input, parsed_inputs)
}
\arguments{
\item{indicator_result}{Result from indicator calculation function.}
\item{input}{List. Shiny input values.}
\item{parsed_inputs}{List. Pre-parsed inputs (breaks, labels, limits, etc.).}
}
\value{
List of parameters ready for plot() function.
}
\examples{
\dontrun{
params <- create_map_plot_params(indicator_result, input, parsed)
plot <- do.call(plot, params)
}
}
