% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_timeseries.R
\name{create_ts_plot}
\alias{create_ts_plot}
\title{Create Time Series Plot}
\description{
Creates a ggplot time series from calculated indicator result.
}
\usage{
create_ts_plot(indicator_ts, params)
}
\arguments{
\item{indicator_ts}{Result from calc_indicator_ts().}
\item{params}{List of plotting parameters.}
}
\value{
ggplot object.
}
\examples{
\dontrun{
ts_result <- calc_indicator_ts(...)
params <- create_ts_plot_params(...)
plot <- create_ts_plot(ts_result, params)
}
}
