% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_timeseries.R
\name{create_ts_plot_params}
\alias{create_ts_plot_params}
\title{Create Time Series Plot Parameters}
\description{
Builds parameter list for time series plotting from UI inputs.
}
\usage{
create_ts_plot_params(input)
}
\arguments{
\item{input}{List. Shiny input values.}
}
\value{
List of parameters for plot() function.
}
\examples{
\dontrun{
params <- create_ts_plot_params(input)
plot <- do.call(plot, c(list(x = ts_result), params))
}
}
