% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_logic.R
\name{get_default_resolution}
\alias{get_default_resolution}
\title{Get Default Resolution}
\description{
Returns the default resolution for a given country type.
}
\usage{
get_default_resolution(countrytype)
}
\arguments{
\item{countrytype}{Character. Type of country data.}
}
\value{
Character. Default resolution ("50" for most, "110" for tiny_countries).
}
\examples{
\dontrun{
get_default_resolution("countries")  # Returns: "50"
get_default_resolution("tiny_countries")  # Returns: "110"
}
}
