% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_logic.R
\name{get_spatial_level_description}
\alias{get_spatial_level_description}
\title{Get Spatial Level Description}
\description{
Returns a human-readable description of a spatial level.
}
\usage{
get_spatial_level_description(spatiallevel)
}
\arguments{
\item{spatiallevel}{Character. Spatial level.}
}
\value{
Character. Description of the spatial level.
}
\examples{
\dontrun{
get_spatial_level_description("country")  # "Countries"
get_spatial_level_description("continent")  # "Continents"
}
}
