% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_viz.R
\name{server_timeseries_plot}
\alias{server_timeseries_plot}
\title{Server Time Series Plot}
\description{
Handles time series plot generation including input validation, parameter
processing, and plot rendering. Creates plot_to_print_ts reactive and
output$plot_ts renderPlot.
}
\usage{
server_timeseries_plot(input, output, session, r, plot_to_render_ts)
}
\arguments{
\item{input}{Shiny input object}
\item{output}{Shiny output object}
\item{session}{Shiny session object}
\item{r}{Reactive values object containing dataCube1 and other shared state}
}
\value{
List containing:
\itemize{
  \item plot_to_print_ts: reactive expression returning the ggplot time series object
  \item render_ts: function to render the time series output
}
}
\examples{
\dontrun{
# In server function:
ts_module <- server_timeseries_plot(input, output, session, r)
}
}
