% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{set_data_cube}
\alias{set_data_cube}
\title{Set Data Cube}
\description{
Stores a processed data cube in the reactive values store.
Automatically creates a copy for filtering.
}
\usage{
set_data_cube(r, cube)
}
\arguments{
\item{r}{ReactiveValues object containing app state.}
\item{cube}{Object of class 'processed_cube'.}
}
\value{
Invisible NULL.
}
\examples{
\dontrun{
cube <- load_data_cube("data.csv")
set_data_cube(r, cube)
}
}
